% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{api_call_cache_directory_name}
\alias{api_call_cache_directory_name}
\title{directory name for where api calls are cached between sessions}
\usage{
api_call_cache_directory_name()
}
\description{
directory name for where api calls are cached between sessions
}
