% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{api_client_file_name}
\alias{api_client_file_name}
\title{file name for where api client is saved between sessions}
\usage{
api_client_file_name()
}
\description{
file name for where api client is saved between sessions
}
