% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{available_surveys}
\alias{available_surveys}
\title{DHS surveys that can be downloaded}
\usage{
available_surveys(your_email, your_password, your_project, output_dir,
  max_urls = NULL)
}
\arguments{
\item{your_email}{Character for email address for DHS website}

\item{your_password}{Character for password for DHS website}

\item{your_project}{Character string for the name of your project that gives you access to the DHS database}

\item{output_dir}{Character for file directory where surveys will be downloaded to when using \code{survey_download}.}

\item{max_urls}{Numeric or maximum number of urls to generate. Default = NULL, which is all}
}
\value{
Returns \code{"data.frame"} of length 3:
\itemize{
      \item{"country"}{Survey Country}
      \item{"year"}{Survey year}
      \item{"Apr_Ctry_list_id"}{Unique numeric for your project for auth purposes}
      \item{"output_folder"}{Folder root directory for where surveys will be downloaded to, based on \strong{output_dir} argument}
      \item{"full_url"}{Full file url for survey download purposes}
      \item{"file_format"}{File format for full_url}
      }
}
\description{
Create a data frame of surveys that your log in can download
}
\note{
Credit for function to \url{https://github.com/ajdamico/lodown/blob/master/R/dhs.R}
}
\examples{
rdhs:::available_surveys(your_email="rdhs.tester@gmail.com",
your_password="rdhstesting",
your_project="Testing Malaria Investigations",
output_dir=tempdir(), max_urls = 1)



}
