% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{client_cache_date}
\alias{client_cache_date}
\title{Pull last cache date}
\usage{
client_cache_date(root)
}
\arguments{
\item{root}{Character for root path to where client,
caches, surveys etc. will be stored.}
}
\description{
Pull last cache date
}
