% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_api_responses}
\alias{collapse_api_responses}
\title{collapse API response list}
\usage{
collapse_api_responses(x)
}
\arguments{
\item{x}{List of lists from API to be collapsed}
}
\description{
collapse API response list
}
