% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datasets.R
\name{data_and_labels}
\alias{data_and_labels}
\title{Create list of dataset and its variable names}
\usage{
data_and_labels(dataset)
}
\arguments{
\item{dataset}{Any read in dataset created by \code{get_datasets}}
}
\description{
Function to give the former output of get_datasets as it can be nice to have both the
definitions and the dataset attached together
}
