% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_labelled.R
\name{delabel_df}
\alias{delabel_df}
\title{convert labelled data frame to data frame of just characters}
\usage{
delabel_df(df)
}
\arguments{
\item{df}{data frame to convert labelled elements of. Likely this will be
the output of \code{\link{extract_dhs}}.}
}
\value{
A data frame of de-labelled elements
}
\description{
convert labelled data frame to data frame of just characters
}
\examples{
df1 <- data.frame(
area = haven::labelled(c(1L, 2L, 3L), c("reg 1"=1,"reg 2"=2,"reg 3"=3)),
climate = haven::labelled(c(0L, 1L, 1L), c("cold"=0,"hot"=1))
)

df_char <- delabel_df(df = df1)

}
