% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{dhs_authenticate}
\alias{dhs_authenticate}
\title{DHS Wesbite Authentication}
\usage{
dhs_authenticate(your_email, your_password, your_project)
}
\arguments{
\item{your_email}{Character for email address for DHS website}

\item{your_password}{Character for password for DHS website}

\item{your_project}{Character string for the name of your
project that gives you access to the DHS database}
}
\value{
Returns list of length 3:
\itemize{
      \item{"user_name"}{your email usually}
      \item{"user_pass"}{your pasword you provided}
      \item{"proj_id"}{your project number that will enable your project to be accessed downstream}
      }
}
\description{
Autheticate Users for DHS website
}
\note{
Credit for function to \url{https://github.com/ajdamico/lodown/blob/master/R/dhs.R}
}
\examples{

dhs_authenticate(your_email="rdhs.tester@gmail.com",your_password="rdhstesting",your_project="Testing Malaria Investigations")



}
