% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\name{dhs_dataUpdates}
\alias{dhs_dataUpdates}
\title{API request of DHS Data Updates}
\usage{
dhs_dataUpdates(client = NULL, all_results = TRUE, lastUpdate = NULL,
  f = NULL, returnFields = NULL, perPage = NULL, page = NULL)
}
\arguments{
\item{client}{If the api request should be cached, then provide a client
object created by \code{\link{client_dhs}}}

\item{all_results}{Boolean for if all results should be returned. If FALSE
then the specifed page only will be returned. Default = TRUE.}

\item{lastUpdate}{Specify a date or Unix time to filter the updates by.
Only results for data that have been updated on or after the specified
date will be returned.}

\item{f}{You can specify the format of the data returned from the query as
HTML, JSON, PJSON, geoJSON, JSONP, XML or CSV. The default data format is
JSON.}

\item{returnFields}{Specify a list of attributes to be returned.}

\item{perPage}{Specify the number of results to be returned per page. By
default the API will return 100 results.}

\item{page}{Allows specifying a page number to obtain for the recordset.
By default the API will return page 1.}
}
\value{
Returns a `data.table` of 9 (or less if `returnFields` is provided)
  indicators or surveys that have been added/updated or removed. A detailed
  description of all the attributes returned is provided at
  \url{https://api.dhsprogram.com/rest/dhs/dataupdates/fields}
}
\description{
API request of DHS Data Updates
}
\examples{
## not run only because they take a very long time and interact with an API
\dontrun{
dat <- dhs_dataUpdates(lastUpdate="20150901",all_results=FALSE)
dat <- dhs_dataUpdates(f="html",all_results=FALSE)
}
}
