% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dhs_gps_data_format}
\alias{dhs_gps_data_format}
\title{DHS GPS Data Format}
\format{A dataframe of 20 observations of 3 variables:

\code{dhs_gps_data_format}: A dataframe of GPS data descriptions.
\itemize{
      \item{"Name"}
      \item{"Type"}
      \item{"Description"}
      }}
\usage{
data(dhs_gps_data_format)
}
\description{
Data frame to describe the data encoded in DHS GPS files
}
\keyword{datasets}
