% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dhs_read_dataset}
\alias{dhs_read_dataset}
\title{raed in dhs standard file types}
\usage{
dhs_read_dataset(file, reformat = TRUE)
}
\arguments{
\item{file}{path to file to be read}

\item{reformat}{boolean detailing if datasets should be nicely reformatted. Default = TRUE}
}
\description{
raed in dhs standard file types
}
