% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\name{dhs_surveyCharacteristics}
\alias{dhs_surveyCharacteristics}
\title{API request of DHS Survey Characteristics}
\usage{
dhs_surveyCharacteristics(countryIds = NULL, indicatorIds = NULL,
  surveyIds = NULL, surveyYear = NULL, surveyYearStart = NULL,
  surveyYearEnd = NULL, surveyType = NULL, f = NULL,
  returnFields = NULL, perPage = NULL, page = NULL, client = NULL,
  force = FALSE, all_results = TRUE)
}
\arguments{
\item{countryIds}{Specify a comma separated list of country ids to filter
by. For a list of countries use
\code{dhs_countries(returnFields=c("CountryName","DHS_CountryCode"))}}

\item{indicatorIds}{Specify a comma separated list of indicators ids to
filter by. For a list of indicators use
\code{dhs_indicators(returnFields=c("IndicatorId","Label","Definition"))}}

\item{surveyIds}{Specify a comma separated list of survey ids to filter by.
For a list of surveys use
\code{dhs_surveys(returnFields=c("SurveyId","SurveyYearLabel",
"SurveyType","CountryName"))}}

\item{surveyYear}{Specify a comma separated list of survey years to
filter by.}

\item{surveyYearStart}{Specify a range of Survey Years to filter Survey
Characteristics on. surveyYearStart is an inclusive value. Can be used
alone or in conjunction with surveyYearEnd.}

\item{surveyYearEnd}{Specify a range of Survey Years to filter Survey
Characteristics on. surveyYearEnd is an inclusive value. Can be used alone
or in conjunction with surveyYearStart.}

\item{surveyType}{Specify a survey type to filter by.}

\item{f}{You can specify the format of the data returned from the query as
HTML, JSON, PJSON, geoJSON, JSONP, XML or CSV. The default data format
is JSON.}

\item{returnFields}{Specify a list of attributes to be returned.}

\item{perPage}{Specify the number of results to be returned per page. By
default the API will return 100 results.}

\item{page}{Allows specifying a page number to obtain for the recordset. By
default the API will return page 1.}

\item{client}{If the api request should be cached, then provide a client
object created by \code{\link{client_dhs}}}

\item{force}{Should we force fetching the API results, and ignore any
cached results we have. Default = FALSE}

\item{all_results}{Boolean for if all results should be returned. If FALSE
then the specifed page only will be returned. Default = TRUE.}
}
\value{
Returns a `data.table` of 2 (or less if `returnFields` is provided)
  survey characteristics. A survey can be labeled with one or more of these
  survey characteristics. A description of all the attributes returned is
  provided at
  \url{https://api.dhsprogram.com/rest/dhs/surveycharacteristics/fields}
}
\description{
API request of DHS Survey Characteristics
}
\examples{
## not run only because they take a very long time and interact with an API
\dontrun{
dat <- dhs_surveyCharacteristics(countryIds="EG",all_results=FALSE)
dat <- dhs_surveyCharacteristics(indicatorIds="FE_FRTR_W_TFR",
all_results=FALSE)
dat <- dhs_surveyCharacteristics(surveyIds="SN2010DHS,all_results=FALSE")
dat <- dhs_surveyCharacteristics(surveyYear="2010,all_results=FALSE")
dat <- dhs_surveyCharacteristics(surveyYearStart="2006",all_results=FALSE)
dat <- dhs_surveyCharacteristics(surveyYearStart="1991",
surveyYearEnd="2006",all_results=FALSE)
dat <- dhs_surveyCharacteristics(surveyType="DHS",all_results=FALSE)
dat <- dhs_surveyCharacteristics(f="html",all_results=FALSE)
}
}
