% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\name{dhs_ui_updates}
\alias{dhs_ui_updates}
\title{API request of DHS UI Updates}
\usage{
dhs_ui_updates(lastUpdate = NULL, f = NULL, returnFields = NULL,
  perPage = NULL, page = NULL, client = NULL, force = FALSE,
  all_results = TRUE)
}
\arguments{
\item{lastUpdate}{Specify a date or Unix time to filter the updates by. Only
results for interfaces that has been updated on or after the specified
date will be returned.}

\item{f}{You can specify the format of the data returned from the query as
HTML, JSON, PJSON, geoJSON, JSONP, XML or CSV. The default data format
is JSON.}

\item{returnFields}{Specify a list of attributes to be returned.}

\item{perPage}{Specify the number of results to be returned per page. By
default the API will return 100 results.}

\item{page}{Allows specifying a page number to obtain for the API request. By
default the API will return page 1.}

\item{client}{If the API request should be cached, then provide a client
object created by \code{\link{client_dhs}}}

\item{force}{Should we force fetching the API results, and ignore any
cached results we have. Default = FALSE}

\item{all_results}{Boolean for if all results should be returned. If FALSE
then the specified page only will be returned. Default = TRUE.}
}
\value{
Returns a \code{data.table} of 3 (or less if \code{returnFields} is provided)
  interfaces that have been added/updated or removed. A detailed description
  of all the attributes returned is provided at
  \url{https://api.dhsprogram.com/rest/dhs/uiupdates/fields}
}
\description{
API request of DHS UI Updates
}
\examples{

# The main use for the ui updates API will be to search for the last time
# there was a change to the UI. For example to return all the
# changes since 2018:

dat <- dhs_ui_updates(lastUpdate="20180101")

# A complete list of examples for how each argument to the ui updates API
# endpoint can be provided is given below, which is a copy of each of
# the examples listed in the API at:

https://api.dhsprogram.com/#/api-uiupdates.cfm

\dontrun{
dat <- dhs_ui_updates(lastUpdate="20150901",all_results=FALSE)
dat <- dhs_ui_updates(f="html",all_results=FALSE)
}
}
