% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{download_datasets}
\alias{download_datasets}
\title{Download surveys specified using output of \code{available_surveys}}
\usage{
download_datasets(your_email, your_password, your_project, desired_survey,
  output_dir_root = NULL, download_option = "both", reformat = TRUE)
}
\arguments{
\item{your_email}{Character for email address for DHS website}

\item{your_password}{Character for password for DHS website}

\item{your_project}{Character string for the name of your project that gives you access to the DHS database}

\item{desired_survey}{Row from \code{available_surveys}}

\item{output_dir_root}{Directory where files are to be downloaded to}

\item{download_option}{Chracter dictating how the durvey is stored when downloaded. Must be one of:
\itemize{
      \item{"zip"} - Just the zip. "z" or anything like will match
      \item{"ex"} - Just the extracted zip. "e" or anything like will match
      \item{"rds"} - Just the read in and saved rds. "r" or anything like will match
      \item{"both"} - Both the rds and extract. "b" or anything like will match
}}

\item{reformat}{Boolean detailing whether dataset rds should be reformatted for ease of use later. Default = TRUE}
}
\description{
Create a data frame of surveys that your log in can download
}
