% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dta_factor_format}
\alias{dta_factor_format}
\title{reformat dta to be more useful}
\usage{
dta_factor_format(dta)
}
\arguments{
\item{dta}{read in dta dataset using  foreign::read.dta(file,convert.dates = FALSE,
convert.factors = FALSE,warn.missing.labels = FALSE)}
}
\value{
list with the eformatted dataset and the code descriptions
}
\description{
reformat dta to be more useful
}
