% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extraction}
\alias{extraction}
\title{DHS survey questions extracted from datasets}
\usage{
extraction(questions, available_datasets, geo_surveys, add_geo = FALSE)
}
\arguments{
\item{questions}{Output of \code{R6_client_dhs$public_methods$survey_questions}}

\item{available_datasets}{Datasets that could be available. Output of \code{R6_client_dhs$public_methods$available_datasets}}

\item{geo_surveys}{Geographic Data Survey file paths.}

\item{add_geo}{Boolean detailing if geographic datasets should be added.}
}
\value{
Returns \code{"list"} of length equal to the number of unique country year combinations within questions.
Each list will then contain further lsts of the surveys collected in that country at that time.
}
\description{
Create a list of survey responses extracted using output of \code{R6_client_dhs$public_methods$survey_questions}
}
