% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_dataset_format}
\alias{file_dataset_format}
\title{Returns what the dataset file ending should be for a given filename}
\usage{
file_dataset_format(file_format)
}
\arguments{
\item{file_format}{FileFormat for a file as taken from the API,
e.g. \code{dhs_datasets(returnFields = "FileFormat")}}
}
\value{
One of "dat","dat","sas7bdat","sav" or "dta"
}
\description{
Returns what the dataset file ending should be for a given filename
}
\examples{
file_format <- "Stata dataset (.dta)"
identical(rdhs:::file_dataset_format(file_format),"dta")

}
