% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_downloaded_datasets}
\alias{get_downloaded_datasets}
\alias{get_downloaded_datasets}
\title{Get Variable Labels
Returns a data.frame of}
\usage{
get_downloaded_datasets()

get_downloaded_datasets()
}
\value{
Data.frame of downlaoded datasets

Data.frame of downlaoded datasets
}
\details{
Returns a \code{data.frame} of the datasets that have been downloaded within this client.
  This could be useful if you are without an interent connection and wish to know which saved
  dataset files in your root diretory correspond to which dataset

Returns a \code{data.frame} of the datasets that have been downloaded within this client.
  This could be useful if you are without an interent connection and wish to know which saved
  dataset files in your root diretory correspond to which dataset
}
