% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datasets.R
\name{get_labels_from_dataset}
\alias{get_labels_from_dataset}
\title{Return variable labels from a dataset}
\usage{
get_labels_from_dataset(data, return_all = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} from which to extract variable labels.}

\item{return_all}{Logical whether to return all variables (\code{TRUE})
or only those with labels.}
}
\value{
A \code{data.frame} consisting of the variable name and labels.
}
\description{
Returns variable labels stored as \code{"label"} attribute.
}
