% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_rdhs_config}
\alias{get_rdhs_config}
\title{Get rdhs config}
\usage{
get_rdhs_config()
}
\value{
A \code{data.frame} containing your rdhs config
}
\description{
Gets the rdhs config being used
}
\details{
Returns the config being used by rdhs at the moment. This will
  either be a `data.frame` with class `rdhs_config` or will be NULL if
  this has not been set up yet
}
