% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_var_labels}
\alias{get_var_labels}
\title{Get Survey Variable Labels}
\usage{
get_var_labels(dataset, return_all = TRUE)
}
\arguments{
\item{dataset}{Can be either the file path to a dataset, the dataset as a
`data.frame` or the filenaes of datasets. See details for more information}

\item{return_all}{Logical whether to return all variables (\code{TRUE})
or only those with labels.}
}
\value{
A \code{data.frame} consisting of the variable name and labels.
}
\description{
Return variable labels from a dataset
}
\details{
Returns variable names and their labels from a dataset.
  You can pass for the `data` argument any of
  the following:
  \itemize{
      \item The file path to a saved dataset. This would be the direct
      output of \code{\link{get_datasets}}
      \item A read in dataset, i.e. produced by using \code{\link{readRDS}}
      to load a dataset from
      a file path produced by \code{\link{get_datasets}}
      \item Dataset filenames. These can be found as one of the returned
      fields from \code{\link{dhs_datasets}}. If these datasets have not been
      downloaded before this will download them for you.
      }
}
