% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{haven_factor_format}
\alias{haven_factor_format}
\title{reformat haven read ins to be more useful}
\usage{
haven_factor_format(res)
}
\arguments{
\item{res}{read in dta dataset using  haven::read_dta(file)}
}
\value{
list with the formatted dataset and the code descriptions
}
\description{
reformat haven read ins to be more useful
}
