% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{last_api_update}
\alias{last_api_update}
\title{Pull last DHS API database update time}
\usage{
last_api_update(timeout = 30)
}
\arguments{
\item{timeout}{Numeric for API timeout. Default = 30}
}
\description{
Pull last DHS API database update time
}
