% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_datasets}
\alias{model_datasets}
\title{DHS model datasets}
\format{A dataframe of 36 observations of 14 variables:

\code{model_datasets}: A dataframe of model datasets
\itemize{
      \item{"FileFormat"}
      \item{"FileSize"}
      \item{"DatasetType"}
      \item{"SurveyNum"}
      \item{"SurveyId"}
      \item{"FileType"}
      \item{"FileDateLastModified"}
      \item{"SurveyYearLabel"}
      \item{"SurveyType"}
      \item{"SurveyYear"}
      \item{"DHS_CountryCode"}
      \item{"FileName"}
      \item{"CountryName"}
      \item{"URLS"}
      }}
\usage{
data(model_datasets)
}
\description{
The model datasets from the DHS website in a `data.frame` that is analogous
to those returned by `get_available_datasets()`
}
\keyword{datasets}
