% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhs_flat.R
\name{parse_dcf}
\alias{parse_dcf}
\title{Parse .DCF dictionary file}
\usage{
parse_dcf(dcf, all_lower = TRUE)
}
\arguments{
\item{dcf}{.DCF file path to parse}

\item{all_lower}{logical indicating whether to convert variable labels to lower case. Defaults to `TRUE`.}
}
\value{
data.frame with metadata and labels as attributes
}
\description{
Parse .DCF file
}
\examples{
mrfl_zip <- tempfile()
download.file(paste0("https://dhsprogram.com/customcf/legacy/data/sample_download_dataset.cfm?",
"Filename=ZZMR61FL.ZIP&Tp=1&Ctry_Code=zz&survey_id=0&doctype=dhs"), mrfl_zip, mode="wb")

dcf <- rdhs:::read_zipdata(mrfl_zip, "\\\\.DCF", readLines)
dct <- rdhs:::parse_dcf(dcf)

}
