% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhs_flat.R
\name{parse_meta}
\alias{parse_meta}
\alias{parse_dcf}
\alias{parse_sps}
\title{Parse dataset metadata}
\usage{
parse_dcf(dcf, all_lower = TRUE)

parse_sps(sps, all_lower = TRUE)
}
\arguments{
\item{dcf}{.DCF file as character vector (e.g. from readLines)}

\item{sps}{.SPS file as character vector (e.g. from readLines)}
}
\value{
data.frame with metadata for parsing fixed-width flat file
}
\description{
Parse fixed-width file metadata
}
\examples{
mrfl_zip <- tempfile()
download.file("https://dhsprogram.com/customcf/legacy/data/sample_download_dataset.cfm?Filename=ZZMR61FL.ZIP&Tp=1&Ctry_Code=zz&survey_id=0&doctype=dhs", mrfl_zip)

dcf <- read_zipdata(mrfl_zip, "\\\\.DCF", readLines)
dct <- parse_dcf(dcf)

sps <- read_zipdata(mrfl_zip, "\\\\.SPS", readLines)
dct <- parse_sps(sps)

}
