% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_list_base}
\alias{rbind_list_base}
\title{implementation of data.tables rbindlist}
\usage{
rbind_list_base(x)
}
\arguments{
\item{x}{List of lists to be converted to a data.frame}
}
\description{
implementation of data.tables rbindlist
}
