% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdhs.R
\docType{package}
\name{rdhs}
\alias{rdhs}
\alias{rdhs-package}
\alias{rdhs-package}
\title{\pkg{rdhs} DHS database through R}
\description{
Provides a API client for the DHS API that allows DHS survey indicators
to be queried. This allows desired indicators to be easily found and passed onto
subsequent functionality within rdhs for downloading datasets and conducting basic
data munging tasks.
}
\author{
\strong{Maintainer}: OJ Watson \email{o.watson15@imperial.ac.uk}

Authors:
\itemize{
  \item Jeff Eaton
}

}
