% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datasets.R
\name{read_dhs_dataset}
\alias{read_dhs_dataset}
\title{read in dhs standard file types}
\usage{
read_dhs_dataset(file, reformat = FALSE, all_lower = TRUE, ...)
}
\arguments{
\item{file}{path to zip file to be read}

\item{reformat}{boolean detailing if datasets should be nicely reformatted. Default = TRUE}

\item{all_lower}{Logical indicating whether all value labels should be lower case. Default to `TRUE`.}

\item{...}{Extra arguments to be}
}
\description{
read in dhs standard file types
}
