% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{response_is_json}
\alias{response_is_json}
\title{checks if the response is json or not by looking at the responses headers}
\usage{
response_is_json(x)
}
\description{
checks if the response is json or not by looking at the responses headers
}
