% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{response_to_json}
\alias{response_to_json}
\title{converts response to json by first converting the response to text}
\usage{
response_to_json(x)
}
\arguments{
\item{x}{A response}
}
\description{
converts response to json by first converting the response to text
}
