% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{set_dhs_credentials}
\alias{set_dhs_credentials}
\title{Set DHS login credentials}
\usage{
set_dhs_credentials(credentials, root = NULL)
}
\arguments{
\item{credentials}{File path to where log in credentials are written.
File format should be (each bullet is a new line):
\itemize{
      \item email=dummy@gmail.com
      \item password=dummypass
      \item project=Dummy Project
      }}

\item{root}{Character for root directory to where client,
  caches, surveys etc. will be stored.
Default = \code{rappdirs::user_cache_dir("rdhs", Sys.info()["user"])}}
}
\value{
Invisibly returns the rdhs package environment client
}
\description{
Set DHS login credentials
}
