% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_special}
\alias{unzip_special}
\title{unzip special that catches for 4GB+}
\usage{
unzip_special(zipfile, files = NULL, overwrite = TRUE,
  junkpaths = FALSE, exdir = ".", unzip = "internal",
  setTimes = FALSE)
}
\arguments{
\item{zipfile}{The pathname of the zip file: tilde expansion (see
    \code{\link{path.expand}}) will be performed.}

\item{files}{A character vector of recorded filepaths to be extracted:
    the default is to extract all files.}

\item{overwrite}{If \code{TRUE}, overwrite existing files (the equivalent
    of \command{unzip -o}), otherwise ignore such files (the equivalent of
    \command{unzip -n}).}

\item{junkpaths}{If \code{TRUE}, use only the basename of the stored
    filepath when extracting.  The equivalent of \command{unzip -j}.}

\item{exdir}{The directory to extract files to (the equivalent of
    \code{unzip -d}).  It will be created if necessary.}

\item{unzip}{The method to be used.  An alternative is to use
    \code{getOption("unzip")}, which on a Unix-alike may be set to the
    path to a \command{unzip} program.}

\item{setTimes}{logical.  For the internal method only, should the
    file times be set based on the times in the zip file?  (NB: this
    applies to included files, not to directories.)}
}
\description{
unzip special that catches for 4GB+
}
