% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_warn_fails}
\alias{unzip_warn_fails}
\title{unzips files without throwing warnings}
\usage{
unzip_warn_fails(...)
}
\arguments{
\item{...}{arguments to pass to \code{unzip}}
}
\description{
unzips files without throwing warnings
}
