% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{available_surveys}
\alias{available_surveys}
\title{DHS surveys that can be downloaded}
\usage{
available_surveys(your_email, your_password, your_project,
  datasets_api_results = NULL, surveys_api_results = NULL)
}
\arguments{
\item{your_email}{Character for email address for DHS website}

\item{your_password}{Character for password for DHS website}

\item{your_project}{Character string for the name of your project that gives you access to the DHS database}

\item{datasets_api_results}{Data.table for the api results for the datasets endpoint. Default = NULL and
generated by default if not declared.}

\item{surveys_api_results}{Data.table for the api results for the surveys endpoint. Default = NULL and
generated by default if not declared.}
}
\value{
Returns \code{"data.frame"} of length 14:
\itemize{
      \item{"FileFormat"}
      \item{"FileSize"}
      \item{"DatasetType"}
      \item{"SurveyNum"}
      \item{"SurveyId"}
      \item{"FileType"}
      \item{"FileDateLastModified"}
      \item{"SurveyYearLabel"}
      \item{"SurveyType"}
      \item{"SurveyYear"}
      \item{"DHS_CountryCode"}
      \item{"FileName"}
      \item{"CountryName"}
      \item{"URLS"}
      }
}
\description{
Create a data frame of surveys that your log in can download
}
\note{
Credit for function to \url{https://github.com/ajdamico/lodown/blob/master/R/dhs.R}
}
