% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{dhs_client}
\alias{dhs_client}
\title{Make a dhs client}
\usage{
dhs_client(api_key = NULL, root = rappdirs::user_cache_dir("rdhs",
  Sys.info()["user"]), ...)
}
\arguments{
\item{api_key}{Character for DHS API KEY}

\item{root}{Character for root directory to where client, caches, surveys etc. will be stored.
Default = \code{rappdirs::user_cache_dir("rdhs",Sys.info()["user"])}}

\item{...}{Passed to \code{R6_dhs_client}}
}
\description{
Make a DHS API client
}
