% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\name{dhs_geometry}
\alias{dhs_geometry}
\title{API request of DHS Geometry}
\usage{
dhs_geometry(client = NULL, all_results = TRUE, countryIds = NULL,
  surveyIds = NULL, surveyYear = NULL, surveyYearStart = NULL,
  surveyYearEnd = NULL, surveyType = NULL, f = NULL,
  returnFields = NULL, perPage = NULL, page = NULL)
}
\arguments{
\item{client}{If the api request should be cached, then provide a client
object created by \code{\link{client_dhs}}}

\item{all_results}{Boolean for if all results should be returned. If FALSE
then the specifed page only will be returned. Default = TRUE.}

\item{countryIds}{Specify a comma separated list of country ids to filter
by. For a list of countries use
\code{dhs_countries(returnFields=c("CountryName","DHS_CountryCode"))}}

\item{surveyIds}{Specify a comma separated list of survey ids to filter by.
For a list of surveys use
\code{dhs_surveys(returnFields=c("SurveyId","SurveyYearLabel",
"SurveyType","CountryName"))}}

\item{surveyYear}{Specify a comma separated list of survey years to
filter by.}

\item{surveyYearStart}{Specify a range of Survey Years to filter Geometry
on. surveyYearStart is an inclusive value. Can be used alone or in
conjunction with surveyYearEnd.}

\item{surveyYearEnd}{Specify a range of Survey Years to filter Geometry on.
surveyYearEnd is an inclusive value. Can be used alone or in conjunction
with surveyYearStart.}

\item{surveyType}{Specify a survey type to filter by.}

\item{f}{You can specify the format of the data returned from the query as
HTML, JSON, PJSON, geoJSON, JSONP, XML or CSV. The default data format
is JSON.}

\item{returnFields}{Specify a list of attributes to be returned.}

\item{perPage}{Specify the number of results to be returned per page. By
default the API will return 100 results.}

\item{page}{Allows specifying a page number to obtain for the recordset.
By default the API will return page 1.}
}
\value{
Returns a `data.table` of 7 (or less if `returnFields` is provided)
  geometry with their corresponding details. A detailed description of all
  the attributes returned is provided at
  \url{https://api.dhsprogram.com/rest/dhs/geometry/fields}
}
\description{
API request of DHS Geometry
}
\examples{
## not run only because they take a very long time and interact with an API
\dontrun{
dat <- dhs_geometry(countryIds="EG",all_results=FALSE)
dat <- dhs_geometry(surveyIds="SN2010DHS",all_results=FALSE)
dat <- dhs_geometry(surveyYear="2010",all_results=FALSE)
dat <- dhs_geometry(surveyYearStart="2006",all_results=FALSE)
dat <- dhs_geometry(surveyYearStart="1991", surveyYearEnd="2006",
all_results=FALSE)
dat <- dhs_geometry(surveyType="DHS",all_results=FALSE)
dat <- dhs_geometry(f="geojson",all_results=FALSE)
}
}
