% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\name{dhs_ui_updates}
\alias{dhs_ui_updates}
\title{API request of DHS UI Updates}
\usage{
dhs_ui_updates(lastUpdate = NULL, f = NULL, returnFields = NULL,
  perPage = NULL, page = NULL, client = NULL, force = FALSE,
  all_results = TRUE)
}
\arguments{
\item{lastUpdate}{Specify a date or Unix time to filter the updates by. Only
results for interfaces that has been updated on or after the sepcified
date will be returned.}

\item{f}{You can specify the format of the data returned from the query as
HTML, JSON, PJSON, geoJSON, JSONP, XML or CSV. The default data format
is JSON.}

\item{returnFields}{Specify a list of attributes to be returned.}

\item{perPage}{Specify the number of results to be returned per page. By
default the API will return 100 results.}

\item{page}{Allows specifying a page number to obtain for the recordset. By
default the API will return page 1.}

\item{client}{If the api request should be cached, then provide a client
object created by \code{\link{client_dhs}}}

\item{force}{Should we force fetching the API results, and ignore any
cached results we have. Default = FALSE}

\item{all_results}{Boolean for if all results should be returned. If FALSE
then the specifed page only will be returned. Default = TRUE.}
}
\value{
Returns a `data.table` of 3 (or less if `returnFields` is provided)
  interfaces that have been added/updated or removed. A detailed description
  of all the attributes returned is provided at
  \url{https://api.dhsprogram.com/rest/dhs/uiupdates/fields}
}
\description{
API request of DHS UI Updates
}
\examples{
## not run only because they take a very long time and interact with an API
\dontrun{
dat <- dhs_ui_updates(lastUpdate="20150901",all_results=FALSE)
dat <- dhs_ui_updates(f="html",all_results=FALSE)
}
}
