% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{download_datasets}
\alias{download_datasets}
\title{Download surveys specified using output of \code{available_surveys}}
\usage{
download_datasets(your_email, your_password, your_project, desired_surveys)
}
\arguments{
\item{your_email}{Character for email address for DHS website}

\item{your_password}{Character for password for DHS website}

\item{your_project}{Character string for the name of your project that gives you access to the DHS database}

\item{desired_surveys}{output or rows from \code{available_surveys}}
}
\description{
Create a data frame of surveys that your log in can download
}
\note{
Credit for function to \url{https://github.com/ajdamico/lodown/blob/master/R/dhs.R}
}
\examples{

dhs_surveys <- rdhs:::available_surveys(your_email="rdhs.tester@gmail.com",
your_password="rdhstesting",
your_project="Testing Malaria Investigations",
output_dir=tempdir(),
max_urls=2)

download_datasets(your_email="rdhs.tester@gmail.com",
your_password="rdhstesting",
your_project="Testing Malaria Investigations",
desired_surveys=dhs_surveys[1,])



}
