% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extraction}
\alias{extraction}
\title{DHS survey questions extracted from surveys}
\usage{
extraction(questions, available_surveys, geo_surveys, add_geo = TRUE)
}
\arguments{
\item{questions}{Output of \code{R6_dhs_client$public_methods$survey_questions}}

\item{available_surveys}{Surveys that could be available. Output of \code{R6_dhs_client$public_methods$available_surveys}}

\item{geo_surveys}{Geographic Data Survey file paths.}

\item{add_geo}{Boolean detailing if geographic datasets should be added.}
}
\value{
Returns \code{"list"} of length equal to the number of unique country year combinations within questions.
Each list will then contain further lsts of the surveys collected in that country at that time.
}
\description{
Create a list of survey responses extracted using output of \code{R6_dhs_client$public_methods$survey_questions}
}
