% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datasets.R
\name{factor_format}
\alias{factor_format}
\title{reformat haven and labelled read ins to have no factors or labels}
\usage{
factor_format(res, reformat = FALSE, all_lower = TRUE)
}
\arguments{
\item{res}{dataset to be formatted}

\item{reformat}{Boolean whether to remove all factors and labels and
just return the unfactored data. Default = FALSE}

\item{all_lower}{Logical indicating whether all value labels
should be lower case. Default to `TRUE`.}
}
\value{
list with the formatted dataset and the code descriptions
}
\description{
reformat haven and labelled read ins to have no factors or labels
}
