% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_downloaded_datasets}
\alias{get_downloaded_datasets}
\title{Get Downloaded Datasets}
\usage{
get_downloaded_datasets()
}
\value{
A \code{data.frame} of downloaded datasets
}
\description{
Detail the datasets that you have already downloaded
}
\details{
Returns a \code{data.frame} of the datasets that have been
  downloaded within this client. This could be useful if you are without
  an internet connection and wish to know which saved
  dataset files in your root directory correspond to which dataset
}
\examples{
# get the model datasets included with the package
model_datasets <- model_datasets

# download one of them
g <- get_datasets(dataset_filenames = model_datasets$FileName[1])

# these will then be stored so that we know what datasets we have downloaded
d <- get_downloaded_datasets()

# which returns a names list of file paths to the datasets
d[1]
}
