% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdhs.R
\docType{package}
\name{rdhs}
\alias{rdhs}
\alias{rdhs-package}
\title{\pkg{rdhs} DHS database through R}
\description{
Provides a client for (1) querying the DHS API for survey indicators
and metadata, (2) identifying surveys and datasets for analysis,
(3) downloading survey datasets from the DHS website, (4) loading
datasets and associate metadata into R, and (5) extracting variables
and combining datasets for pooled analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropensci.github.io/rdhs/}
  \item Report bugs at \url{https://github.com/ropensci/rdhs/issues}
}

}
\author{
\strong{Maintainer}: OJ Watson \email{o.watson15@imperial.ac.uk}

Authors:
\itemize{
  \item Jeff Eaton
}

}
